increase entity size from 4 to 8 bytes

+++ src/entity_type.zig
@@ -1,4 +1,4 @@
-pub const EntityId = u32;
+pub const EntityId = u64;
 pub const Entity = packed struct {
     id: EntityId,
 };
